Add delayed queue processing
